Refactor recent portions of vc.py into MSVC module
