+++ SCons/Tool/MSCommon/MSVC/Exceptions.py
@@ -22,40 +22,18 @@
 # WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 """
-Common functions for Microsoft Visual Studio and Visual C/C++.
+Exceptions for Microsoft Visual C/C++.
 """
 
+class VisualCException(Exception):
+    pass
 
-import SCons.Errors
-import SCons.Platform.win32
-import SCons.Util
+class MSVCVersionNotFound(VisualCException):
+    pass
 
-from SCons.Tool.MSCommon.sdk import mssdk_exists, mssdk_setup_env
+class MSVCInternalError(VisualCException):
+    pass
 
-from SCons.Tool.MSCommon.MSVC import (
-    set_msvc_notfound_policy,
-    get_msvc_notfound_policy,
-)
+class MSVCArgumentError(VisualCException):
+    pass
 
-from SCons.Tool.MSCommon.vc import (
-    msvc_exists,
-    msvc_setup_env_tool,
-    msvc_setup_env_once,
-    msvc_version_to_maj_min,
-    msvc_find_vswhere,
-    get_msvc_sdk_versions,
-)
-
-from SCons.Tool.MSCommon.vs import (
-    get_default_version,
-    get_vs_by_version,
-    merge_default_version,
-    msvs_exists,
-    query_versions,
-)
-
-# Local Variables:
-# tab-width:4
-# indent-tabs-mode:nil
-# End:
-# vim: set expandtab tabstop=4 shiftwidth=4:
